+++ packages/retail-ui/components/Fias/models/Address.ts
@@ -1,3 +1,4 @@
+import { FiasLocale } from '../locale';
 import {
   AddressErrors,
   AddressFields,
@@ -9,7 +10,6 @@ import {
   FiasValue,
   Fields,
   ExtraFields,
-  FiasLocale,
   FieldsSettings,
   FiasCountry,
 } from '../types';
@@ -61,7 +61,7 @@ export class Address {
     Fields.planningstructure,
   ];
 
-  public static ALL_PARENTS_SEARCH_FIELDS = [
+  public static NOT_ONLY_DIRECT_PARENT_SEARCH_FIELDS = [
     Fields.district,
     Fields.city,
     Fields.intracityarea,
@@ -156,7 +156,7 @@ export class Address {
         }
 
         if (!address.isAllowedToFill(field)) {
-          error = locale[`${field}FillBefore`];
+          error = locale[`${field}FillBefore` as keyof FiasLocale];
         }
 
         if (Boolean(error)) {
@@ -199,6 +199,8 @@ export class Address {
     return verifiedAddress;
   };
 
+  // TODO: Hide invisible fields without removing them.
+  // (removing can break the verification)
   public static filterVisibleFields = (
     fields: { [key in Fields]?: any },
     fieldsSettings: FieldsSettings,
@@ -345,8 +347,10 @@ export class Address {
     return false;
   };
 
-  public isAllowedToSearchThroughAllParents = (field?: Fields): boolean => {
-    return Boolean(field && Address.ALL_PARENTS_SEARCH_FIELDS.includes(field));
+  // doesn't work on api side yet
+  // @see https://yt.skbkontur.ru/issue/PS-1401
+  public isAllowedToSearchThroughChildrenOfDirectParent = (field?: Fields): boolean => {
+    return Boolean(field && Address.NOT_ONLY_DIRECT_PARENT_SEARCH_FIELDS.includes(field));
   };
 
   public hasOnlyIndirectParent = (field?: Fields): boolean => {