+++ packages/retail-ui/components/Tooltip/Tooltip.tsx
@@ -34,6 +34,8 @@ export type TooltipTrigger =
   | 'click'
   /** Фокус на children */
   | 'focus'
+  /** Наведение на children и на тултип и фокус на children */
+  | 'hover&focus'
   /** Просто открыт */
   | 'opened'
   /** Просто закрыт */
@@ -70,6 +72,19 @@ export interface TooltipProps {
 
   pos: PopupPosition;
 
+  /**
+   * Триггер открытия тултипа
+   * ```ts
+   * type TooltipTrigger =
+   * | 'hover'
+   * | 'click'
+   * | 'focus'
+   * | 'hover&focus'
+   * | 'opened'
+   * | 'closed'
+   * | 'hoverAnchor';
+   * ```
+   */
   trigger: TooltipTrigger;
 
   /**
@@ -123,6 +138,7 @@ export interface TooltipProps {
 
 export interface TooltipState {
   opened: boolean;
+  focused: boolean;
 }
 
 class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
@@ -151,9 +167,9 @@ class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
 
   public static closeDelay = 100;
 
-  private static triggersWithoutCloseButton: TooltipTrigger[] = ['hover', 'hoverAnchor', 'focus'];
+  private static triggersWithoutCloseButton: TooltipTrigger[] = ['hover', 'hoverAnchor', 'focus', 'hover&focus'];
 
-  public state: TooltipState = { opened: false };
+  public state: TooltipState = { opened: false, focused: false };
 
   private hoverTimeout: Nullable<number> = null;
   private contentElement: Nullable<HTMLElement> = null;
@@ -314,6 +330,17 @@ class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
           },
         };
 
+      case 'hover&focus':
+        return {
+          popupProps: {
+            onFocus: this.handleFocus,
+            onBlur: this.handleBlur,
+            onMouseEnter: this.handleMouseEnter,
+            onMouseLeave: this.handleMouseLeave,
+            useWrapper,
+          },
+        };
+
       default:
         throw new Error('Unknown trigger specified: ' + props.trigger);
     }
@@ -341,8 +368,11 @@ class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
   };
 
   private handleMouseLeave = (event: MouseEventType) => {
-    const triggerIsHover = this.props.trigger === 'hover';
-    if (triggerIsHover && event.relatedTarget === this.contentElement) {
+    if (this.props.trigger === 'hover&focus' && this.state.focused) {
+      return;
+    }
+
+    if (this.props.trigger === 'hover' && event.relatedTarget === this.contentElement) {
       return;
     }
 
@@ -380,11 +410,13 @@ class Tooltip extends React.PureComponent<TooltipProps, TooltipState> {
   }
 
   private handleFocus = () => {
+    this.setState({ focused: true });
     this.open();
   };
 
   private handleBlur = () => {
     this.close();
+    this.setState({ focused: false });
   };
 
   private handleCloseButtonClick = (event: React.MouseEvent<HTMLElement>) => {