+++ packages/retail-ui/components/TopBar/locale/index.ts
@@ -1,5 +1,5 @@
 import { LocaleHelper } from '../../LocaleProvider/LocaleHelper';
-import en_EN from './locales/en';
+import en_GB from './locales/en';
 import ru_RU from './locales/ru';
 import { TopBarLocale } from './types';
 
@@ -7,5 +7,5 @@ export * from './types';
 
 export const TopBarLocaleHelper = new LocaleHelper<TopBarLocale>({
   ru_RU,
-  en_EN,
+  en_GB,
 });