+++ packages/retail-ui/lib/events/tabListener.ts
@@ -1,8 +1,9 @@
-const TAB = 9;
+const KEYCODE_TAB = 9;
+
 class TabListener {
   public isTabPressed: boolean = false;
   constructor() {
-    window.addEventListener('keydown', event => (this.isTabPressed = event.keyCode === TAB));
+    window.addEventListener('keydown', event => (this.isTabPressed = event.keyCode === KEYCODE_TAB));
     window.addEventListener('mousedown', () => (this.isTabPressed = false));
   }
 }